添加了一个server向meta_server的注册的示例 整合了事务模块 日志恢复模块

+++ include/storage/LogStorage.h
@@ -58,7 +58,7 @@ public:
 
     bool ReadLog(char *log_data, int size, size_t offset) {
 
-        if (offset >= GetFileSize(log_name_)) {
+        if (offset >= (size_t)GetFileSize(log_name_)) {
             return false;
         }
